Move implementations of add_metaclass and with_metaclass to their own module.
The overlay implementations of six.add_metaclass and six.with_metaclass are
independent of six. Moving them to their own module will let them be used by
other libraries implementing the same functionality (e.g. `future`)

PiperOrigin-RevId: 216763783
